requirements.txt
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
pyserial==3.5
python-socketio==5.10.0
psycopg2-binary==2.9.9
pywebview==4.4.1
